Init minimal blog post viewer
This is pre-alpha.0001-RC1 quality code, but it's late and it's working.

Loads a list of posts from a post meta JSON file produced by Jekyll and
then renders links to view those posts. When a post view loads, it
doesn't have the post content so it fetches it from individual post
content files that Jekyll also produces, then it renders the HTML.

Pretty cool!