Allow live editing in development.
This commit adds two new elements to the project.
	- It serves the generated JavaScript from `webpack-dev-server`
	- It incorporates `react-hot-loader` to enable live editing

**webpack** sits somewhat awkwardly here because we still need
**Jekyll** to be running and serving up our files. This is because if
we don't start Jekyll, it won't produce and serve the `json` files we need.

On the other hand, the motivating factor for moving away from Jekyll's
dev server is that it was super slow to respond to file changes. In
order to facilitate this, we create a new static HTML file that pulls
from another port, the one on which we start **webpack-dev-server**.
This HTML file won't change and thus won't trigger a Jekyll rebuild
after every JavaScript update, but the webpack server can still monitor
all of its dependent files.